requirements.txt
py-cord==2.6.1
google-generativeai>=0.7.0
python-dotenv>=1.0.0
